refactor(chat): split compaction_run into per-strategy submodules
Pure code movement — zero behaviour change. The 1,655-line
`compaction_run.rs` was review-hostile; splitting into a module
directory makes each strategy (and its tests) a separate 100–460 line
file.

New layout:

  crates/chat/src/compaction_run/
  ├── mod.rs                  (266 LOC) — CompactionRunError,
  │                                       run_compaction() dispatcher,
  │                                       extract_summary_body(), plus
  │                                       dispatcher + helper tests.
  ├── shared.rs               (460 LOC) — HeadTailBounds, compute_boundaries,
  │                                       finalize_kept, in_place_prune_or_err,
  │                                       message_tokens, is_tool_role_value,
  │                                       align_boundary_forward_past_tool_group,
  │                                       prune_tool_results_before,
  │                                       prune_single_tool_result,
  │                                       sanitize_tool_pairs, stub_tool_result,
  │                                       build_middle_marker,
  │                                       build_summary_message.
  ├── deterministic.rs        ( 76 LOC) — run() + 1 dispatch test.
  ├── recency_preserving.rs   (300 LOC) — run() + 5 tests.
  ├── structured.rs           (441 LOC) — STRUCTURED_TEMPLATE +
  │                                       STRUCTURED_SYSTEM_INSTRUCTIONS +
  │                                       first/iterative instruction
  │                                       builders + extract_previous_summary
  │                                       + run() + 6 tests.
  ├── llm_replace.rs          (131 LOC) — run() + 2 tests.
  └── test_support.rs         (149 LOC) — #[cfg(test)]-only StubProvider
                                          shared between the LLM-backed
                                          strategies' test suites.

Visibility: shared helpers are `pub(super)` so only siblings in
compaction_run/** can see them. Strategy modules call each other via
`super::strategy::run` (e.g. structured's fallback to
recency_preserving::run). The public API is unchanged —
`run_compaction`, `extract_summary_body`, and `CompactionRunError`
remain the only items crossing the `pub(crate)` boundary, so
crates/chat/src/lib.rs needs no changes.

Feature gating is cleaner too: `mod structured;` and `mod llm_replace;`
are gated with `#[cfg(feature = "llm-compaction")]` at the mod
declaration instead of scattered `#[cfg]` around bodies inside a
single file. The StubProvider in test_support.rs is always compiled in
test builds so both feature configurations can use it, with
`#[allow(dead_code)]` to cover the helpers that are only exercised
when the feature is on.

Tests now group by strategy in the test runner output:
  compaction_run::deterministic::tests::…
  compaction_run::recency_preserving::tests::…
  compaction_run::structured::tests::…
  compaction_run::llm_replace::tests::…
  compaction_run::tests::…  (dispatcher-level: empty_history,
                             extract_summary_body, feature-off gates)

Validation:
- cargo +nightly-2025-11-30 fmt --all -- --check — clean
- cargo +nightly-2025-11-30 clippy -p moltis-chat -p moltis-config --all-targets -- -D warnings — clean
- cargo +nightly-2025-11-30 clippy -p moltis-chat --no-default-features --all-targets -- -D warnings — clean
- cargo test -p moltis-chat --lib — 219 passed (unchanged)
- cargo test -p moltis-config --lib — 185 passed (unchanged)
- cargo check -p moltis-chat --no-default-features --tests — clean

Each strategy file is now small enough to review on its own, and new
modes in future (e.g. when moltis-8me adds the auxiliary-model
subsystem) land as edits to a single file rather than scrolling
through a monolith.

Entire-Checkpoint: 2b0c022927ce
